Move air transport content to separate function
